ref(seer): Move seerAgentHooks into sentry, to make it easier to create stories later
